Fix disabled autolock for button free screen
